// front/src/app/models/usuario.ts
export interface Usuario {
    id?: number;
    nombre: string;
    username: string;
    apellidos: string;
    email: string;
    telefono: string;
    provincia: string;
    localidad: string;
    direccion: string;
    fechaAlta: string;
    fechaNacimiento: string;
    fotoPerfil: string;
    password: string;
    rol: string;
    sexo: string;
  }